刘永红：增加了一个测试类TestStudentDao.java
